fix(web): percent-encode object keys in v3 table filter URL serialization
getCommaArrayParam wrote the key slot of the column;type;key;operator;value
filter URL format raw, so a metadata/score key containing ';' shifted the
fields on decode and the filter was silently dropped on reload or when
sharing a link. Percent-encode the key on encode and decode it tolerantly
(raw fallback keeps legacy unencoded URLs working). Extract the pure
encoder/decoder from useFilterState.ts for unit tests.
